Improvements to `.R` file chunks:
- Add option Run Chunk and Move to Next Chunk
- Highlight code chunks differently for base `.R` file.
Highlighting the background just highlights the
whole entire screen.
Instead match `.py` and use border for current cell
